move plugin base methods to plugins.__init__
