+++ pidtree_bcc/plugins/loginuidmap.py
@@ -2,7 +2,7 @@ import pwd
 import sys
 from typing import Tuple
 
-from pidtree_bcc.plugin import BasePlugin
+from pidtree_bcc.plugins import BasePlugin
 
 
 class Loginuidmap(BasePlugin):
